+++ pkg/store/bucket.go
@@ -296,9 +296,6 @@ type BucketStore struct {
 
 	// Enables hints in the Series() response.
 	enableSeriesResponseHints bool
-
-	// respPool is a sync.Pool for marshaling Series() responses.
-	respPool sync.Pool
 }
 
 type noopCache struct{}
@@ -399,7 +396,6 @@ func NewBucketStore(
 		enableCompatibilityLabel:    enableCompatibilityLabel,
 		postingOffsetsInMemSampling: postingOffsetsInMemSampling,
 		enableSeriesResponseHints:   enableSeriesResponseHints,
-		respPool:                    sync.Pool{},
 	}
 
 	for _, option := range options {
@@ -1097,17 +1093,8 @@ func (s *BucketStore) Series(req *storepb.SeriesRequest, srv storepb.Store_Serie
 		s.metrics.seriesGetAllDuration.Observe(stats.getAllDuration.Seconds())
 		s.metrics.seriesBlocksQueried.Observe(float64(stats.blocksQueried))
 	}
-
-	var resp *storepb.SeriesResponse
-	defer func(r **storepb.SeriesResponse) {
-		if *r != nil {
-			(*r).Close()
-		}
-	}(&resp)
-
 	// Merge the sub-results from each selected block.
 	tracing.DoInSpan(ctx, "bucket_store_merge_all", func(ctx context.Context) {
-
 		begin := time.Now()
 
 		// NOTE: We "carefully" assume series and chunks are sorted within each SeriesSet. This should be guaranteed by
@@ -1128,15 +1115,7 @@ func (s *BucketStore) Series(req *storepb.SeriesRequest, srv storepb.Store_Serie
 				s.metrics.chunkSizeBytes.Observe(float64(chunksSize(series.Chunks)))
 			}
 			series.Labels = labelpb.ZLabelsFromPromLabels(lset)
-
-			if resp == nil {
-				resp = storepb.NewSeriesResponseWithPool(&series, &s.respPool)
-			} else {
-				resp.Result = &storepb.SeriesResponse_Series{
-					Series: &series,
-				}
-			}
-			if err = srv.Send(resp); err != nil {
+			if err = srv.Send(storepb.NewSeriesResponse(&series)); err != nil {
 				err = status.Error(codes.Unknown, errors.Wrap(err, "send series response").Error())
 				return
 			}
@@ -1159,15 +1138,7 @@ func (s *BucketStore) Series(req *storepb.SeriesRequest, srv storepb.Store_Serie
 			return
 		}
 
-		if resp == nil {
-			resp = storepb.NewHintsSeriesResponseWithPool(anyHints, &s.respPool)
-		} else {
-			resp.Result = &storepb.SeriesResponse_Hints{
-				Hints: anyHints,
-			}
-		}
-
-		if err = srv.Send(resp); err != nil {
+		if err = srv.Send(storepb.NewHintsSeriesResponse(anyHints)); err != nil {
 			err = status.Error(codes.Unknown, errors.Wrap(err, "send series response hints").Error())
 			return
 		}